refactor: group Drive sources into sources/drive/ and Claude parsers into parsers/claude/
Move Drive modules (drive_auth, drive_gateway, drive_source*, drive_types) into
polylogue/sources/drive/ and Claude parser modules (claude_ai_parser,
claude_code_parser, etc.) into polylogue/sources/parsers/claude/.

drive.py → drive/__init__.py and claude.py → claude/__init__.py preserve
the public import surface.

Ref #403
