+++ polylogue/sources/parsers/claude/code_detection.py
