+++ tests/infra/drive_mocks.py
@@ -9,8 +9,8 @@ from datetime import datetime
 from typing import ParamSpec, Protocol, TypeVar
 
 from polylogue.lib.json import json_document, json_document_list
-from polylogue.sources.drive_gateway import DriveListFilesResponse, DrivePayloadRecord
-from polylogue.sources.drive_types import DriveError, DriveNotFoundError
+from polylogue.sources.drive.gateway import DriveListFilesResponse, DrivePayloadRecord
+from polylogue.sources.drive.types import DriveError, DriveNotFoundError
 
 P = ParamSpec("P")
 T = TypeVar("T")
